refactor: :art: improve structure code / changing by js to TS with initial types(need improve more this)
