--- lib/query_util.ts
@@ -1,7 +1,0 @@
-class QueryUtil {
-  static querystring(object: any) {
-    return Object.keys(object).map(key => `${key}=${object[key]}`).join('&');
-  }
-}
-
-export default QueryUtil;
