+++ src/index.ts
@@ -1,0 +1,8 @@
+import {Instagram} from './resources/services/index.service';
+import {PostsService} from "./resources/services/posts/post.service";
+
+export {
+    Instagram,
+    PostsService
+};
+export default Instagram;
