+++ src/resources/services/posts/post.type.d.ts
@@ -1,0 +1,3 @@
+export type IGetPostOptions = {
+    useGraphQL?: boolean ;
+}
