+++ src/util/api_util.ts
@@ -1,18 +1,18 @@
-import https from 'https';
-import QueryUtil from "./query_util";
-const insta = 'https://www.instagram.com/';
+import https = require('https');
+import {INSTAGRAM_BASE_URL} from "../shared/constants";
+import {queryString} from "../shared/helpers/queryString";
 
 class ApiUtil {
-  static get(
-    path: string,
-    sessionId: string,
-    tryParse: boolean = true,
+  public static async get(
+    path: string | any,
+    sessionId?: string | any,
+    tryParse = true,
     params?: any
-  ): any {
+  ): Promise<any> {
     return new Promise((resolve, reject) => {
       params = JSON.stringify({ __a: sessionId ? '1' : undefined, ...params });
-      const url = insta + path + ((params !== '{}')
-        ? ('/?' + QueryUtil.querystring(JSON.parse(params)))
+      const url = INSTAGRAM_BASE_URL + path + ((params !== '{}')
+        ? ('/?' + queryString(JSON.parse(params)))
         : (tryParse ? '/' : ''));
 
       https.get(url, {
@@ -21,18 +21,18 @@ class ApiUtil {
         }
       }, (res: any) => {
         let body = '';
-        res.on('data', chunk => body += chunk);
+        res.on('data', (chunk: any) => body += chunk);
         res.on('end', () => {
           if (res.statusCode !== 200) {
             switch (res.statusCode) {
               case 302: {
                 switch (res.headers.location) {
-                  case insta + 'accounts/login/':
+                  case INSTAGRAM_BASE_URL + 'accounts/login/':
                     return reject(429);
-                  case insta + 'accounts/login/?next=/accounts/edit/%3F__a%3D1':
+                  case INSTAGRAM_BASE_URL + 'accounts/login/?next=/accounts/edit/%3F__a%3D1':
                     return reject(401);
                   default: {
-                    if (res.headers.location.startsWith(insta + 'challenge/?next='))
+                    if (res.headers.location.startsWith(INSTAGRAM_BASE_URL + 'challenge/?next='))
                       return reject(409);
                     reject(res.statusCode);
                   }
@@ -47,7 +47,10 @@ class ApiUtil {
             }
             catch (_) {
               try {
-                resolve(Object.values(Object.values(JSON.parse(body.match(/_sharedData = (.+);/)[1])['entry_data'])[0][0]['graphql'])[0]);
+                // @ts-ignore
+                resolve(Object.values(Object.values(
+                    // @ts-ignore
+                    JSON.parse(body.match(/_sharedData = (.+);/)[1])['entry_data'])[0][0]['graphql'])[0]);
               }
               catch (_) {
                 reject(406);
@@ -63,8 +66,8 @@ class ApiUtil {
     });
   }
 
-  static graphQL(query, queryHash, sessionId) {
-    return ApiUtil.get('graphql/query', sessionId, undefined, {
+  public static async graphQL(query: object, queryHash: any, sessionId: any): Promise<any> {
+    return await ApiUtil.get('graphql/query', sessionId, undefined, {
       query_hash: queryHash,
       variables: query ? JSON.stringify(query) : undefined
     });
